--- web/src/app/lp/components/Icon.tsx
@@ -1,183 +1,0 @@
-import type { ReactElement } from "react";
-
-type IconName =
-  | "plus"
-  | "send"
-  | "map-pin"
-  | "train"
-  | "home"
-  | "message-circle"
-  | "cloud"
-  | "sparkles"
-  | "help-circle"
-  | "trash"
-  | "leaf"
-  | "utensils"
-  | "mountain"
-  | "qr-code"
-  | "chevron-left"
-  | "chevron-right"
-  | "clipboard"
-  | "thumbs-up"
-  | "mail"
-  | "arrow-right"
-  | "paintbrush";
-
-const PATHS: Record<IconName, ReactElement> = {
-  plus: <path d="M12 5v14M5 12h14" />,
-  send: <path d="m5 12 14-7-7 14-2-5-5-2z" />,
-  "map-pin": (
-    <>
-      <path d="M20 10c0 6-8 12-8 12s-8-6-8-12a8 8 0 0 1 16 0z" />
-      <circle cx="12" cy="10" r="3" />
-    </>
-  ),
-  train: (
-    <>
-      <rect x="4" y="3" width="16" height="16" rx="2" />
-      <path d="M4 11h16" />
-      <path d="m8 19-2 3" />
-      <path d="m18 22-2-3" />
-      <circle cx="8" cy="15" r=".5" fill="currentColor" />
-      <circle cx="16" cy="15" r=".5" fill="currentColor" />
-    </>
-  ),
-  home: (
-    <>
-      <path d="m3 9 9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z" />
-      <path d="M9 22V12h6v10" />
-    </>
-  ),
-  "message-circle": (
-    <path d="M21 11.5a8.38 8.38 0 0 1-.9 3.8 8.5 8.5 0 0 1-7.6 4.7 8.38 8.38 0 0 1-3.8-.9L3 21l1.9-5.7a8.38 8.38 0 0 1-.9-3.8 8.5 8.5 0 0 1 4.7-7.6 8.38 8.38 0 0 1 3.8-.9h.5a8.48 8.48 0 0 1 8 8v.5z" />
-  ),
-  cloud: <path d="M17.5 19H9a7 7 0 1 1 6.71-9h1.79a4.5 4.5 0 1 1 0 9Z" />,
-  sparkles: (
-    <>
-      <path d="m12 3-1.9 5.8a2 2 0 0 1-1.3 1.3L3 12l5.8 1.9a2 2 0 0 1 1.3 1.3L12 21l1.9-5.8a2 2 0 0 1 1.3-1.3L21 12l-5.8-1.9a2 2 0 0 1-1.3-1.3z" />
-      <path d="M5 3v4" />
-      <path d="M19 17v4" />
-      <path d="M3 5h4" />
-      <path d="M17 19h4" />
-    </>
-  ),
-  "help-circle": (
-    <>
-      <circle cx="12" cy="12" r="10" />
-      <path d="M9.1 9a3 3 0 0 1 5.8 1c0 2-3 3-3 3" />
-      <path d="M12 17h.01" />
-    </>
-  ),
-  trash: (
-    <>
-      <path d="M3 6h18" />
-      <path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6" />
-      <path d="M8 6V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2" />
-    </>
-  ),
-  leaf: (
-    <>
-      <path d="M11 20A7 7 0 0 1 9.8 6.1C15.5 5 17 4.48 19 2c1 2 2 4.18 2 8 0 5.5-4.78 10-10 10Z" />
-      <path d="M2 21c0-3 1.85-5.36 5.08-6" />
-    </>
-  ),
-  utensils: (
-    <>
-      <path d="M3 2v7c0 1.1.9 2 2 2h4a2 2 0 0 0 2-2V2" />
-      <path d="M7 2v20" />
-      <path d="M21 15V2a5 5 0 0 0-5 5v6c0 1.1.9 2 2 2h3Zm0 0v7" />
-    </>
-  ),
-  mountain: <path d="m8 3 4 8 5-5 5 15H2L8 3z" />,
-  "qr-code": (
-    <>
-      <rect width="5" height="5" x="3" y="3" rx="1" />
-      <rect width="5" height="5" x="16" y="3" rx="1" />
-      <rect width="5" height="5" x="3" y="16" rx="1" />
-      <path d="M21 16h-3a2 2 0 0 0-2 2v3" />
-      <path d="M21 21v.01" />
-      <path d="M12 7v3a2 2 0 0 1-2 2H7" />
-      <path d="M3 12h.01" />
-      <path d="M12 3h.01" />
-      <path d="M12 16v.01" />
-      <path d="M16 12h1" />
-      <path d="M21 12v.01" />
-      <path d="M12 21v-1" />
-    </>
-  ),
-  "chevron-left": <path d="m15 18-6-6 6-6" />,
-  "chevron-right": <path d="m9 18 6-6-6-6" />,
-  clipboard: (
-    <>
-      <rect width="8" height="4" x="8" y="2" rx="1" ry="1" />
-      <path d="M16 4h2a2 2 0 0 1 2 2v14a2 2 0 0 1-2 2H6a2 2 0 0 1-2-2V6a2 2 0 0 1 2-2h2" />
-    </>
-  ),
-  "thumbs-up": (
-    <>
-      <path d="M7 10v12" />
-      <path d="M15 5.88 14 10h5.83a2 2 0 0 1 1.92 2.56l-2.33 8A2 2 0 0 1 17.5 22H7V10l4-8h1a2 2 0 0 1 2 2v.88Z" />
-    </>
-  ),
-  mail: (
-    <>
-      <rect width="20" height="16" x="2" y="4" rx="2" />
-      <path d="m22 7-8.97 5.7a1.94 1.94 0 0 1-2.06 0L2 7" />
-    </>
-  ),
-  "arrow-right": (
-    <>
-      <path d="M5 12h14" />
-      <path d="m12 5 7 7-7 7" />
-    </>
-  ),
-  paintbrush: (
-    <>
-      <path d="m14.622 17.897-10.68-2.913" />
-      <path d="M18.376 2.622a1 1 0 1 1 3.002 3.002L17.36 9.643a.5.5 0 0 0 0 .707l.944.944a2.41 2.41 0 0 1 0 3.408l-.944.944a.5.5 0 0 1-.707 0L8.354 7.348a.5.5 0 0 1 0-.707l.944-.944a2.41 2.41 0 0 1 3.408 0l.944.944a.5.5 0 0 0 .707 0z" />
-      <path d="M9 8c-1.804 2.71-3.97 3.46-6.583 3.948a.507.507 0 0 0-.302.819l7.32 8.883a1 1 0 0 0 1.185.204C12.735 20.405 16 16.792 16 15" />
-    </>
-  ),
-};
-
-type Props = {
-  name: IconName;
-  size?: number;
-  className?: string;
-};
-
-export const Icon = ({ name, size = 16, className }: Props) => (
-  <svg
-    width={size}
-    height={size}
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="1.75"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-    className={className}
-    aria-hidden="true"
-  >
-    {PATHS[name]}
-  </svg>
-);
-
-export const LineIcon = ({
-  size = 16,
-  className,
-}: {
-  size?: number;
-  className?: string;
-}) => (
-  <svg
-    width={size}
-    height={size}
-    viewBox="0 0 24 24"
-    fill="currentColor"
-    className={className}
-    aria-hidden="true"
-  >
-    <path d="M19.365 9.863c.349 0 .63.285.63.631 0 .345-.281.63-.63.63H17.61v1.125h1.755c.349 0 .63.283.63.63 0 .344-.281.629-.63.629h-2.386a.633.633 0 0 1-.628-.629V8.108c0-.345.282-.63.63-.63h2.386c.346 0 .63.285.63.63 0 .349-.284.63-.63.63H17.61v1.125h1.755zm-3.855 3.016a.63.63 0 0 1-.631.629.618.618 0 0 1-.51-.252l-2.443-3.317v2.94a.63.63 0 1 1-1.257 0V8.108a.627.627 0 0 1 .628-.628c.185 0 .406.085.519.232l2.442 3.319V8.108c0-.345.282-.63.63-.63.345 0 .628.285.628.63v4.771zm-5.741 0a.63.63 0 1 1-1.259 0V8.108a.628.628 0 0 1 1.259 0v4.771zm-2.466.629H4.917a.634.634 0 0 1-.63-.629V8.108c0-.345.285-.63.63-.63.348 0 .63.285.63.63v4.141h1.756c.348 0 .629.283.629.63 0 .344-.282.629-.629.629M24 10.314C24 4.943 18.615.572 12 .572S0 4.943 0 10.314c0 4.811 4.27 8.842 10.035 9.608.391.082.923.258 1.058.59.12.301.079.766.038 1.08l-.164 1.02c-.045.301-.24 1.186 1.049.645 1.291-.539 6.916-4.078 9.436-6.975C23.176 14.393 24 12.458 24 10.314" />
-  </svg>
-);
