+++ web/src/app/lp/components/LineIcon.tsx
@@ -1,0 +1,18 @@
+type Props = {
+  size?: number;
+  className?: string;
+};
+
+/** LINE 公式ロゴ。lucide-react に該当アイコンが無いので独自に保持する */
+export const LineIcon = ({ size = 16, className }: Props) => (
+  <svg
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    className={className}
+    aria-hidden="true"
+  >
+    <path d="M19.365 9.863c.349 0 .63.285.63.631 0 .345-.281.63-.63.63H17.61v1.125h1.755c.349 0 .63.283.63.63 0 .344-.281.629-.63.629h-2.386a.633.633 0 0 1-.628-.629V8.108c0-.345.282-.63.63-.63h2.386c.346 0 .63.285.63.63 0 .349-.284.63-.63.63H17.61v1.125h1.755zm-3.855 3.016a.63.63 0 0 1-.631.629.618.618 0 0 1-.51-.252l-2.443-3.317v2.94a.63.63 0 1 1-1.257 0V8.108a.627.627 0 0 1 .628-.628c.185 0 .406.085.519.232l2.442 3.319V8.108c0-.345.282-.63.63-.63.345 0 .628.285.628.63v4.771zm-5.741 0a.63.63 0 1 1-1.259 0V8.108a.628.628 0 0 1 1.259 0v4.771zm-2.466.629H4.917a.634.634 0 0 1-.63-.629V8.108c0-.345.285-.63.63-.63.348 0 .63.285.63.63v4.141h1.756c.348 0 .629.283.629.63 0 .344-.282.629-.629.629M24 10.314C24 4.943 18.615.572 12 .572S0 4.943 0 10.314c0 4.811 4.27 8.842 10.035 9.608.391.082.923.258 1.058.59.12.301.079.766.038 1.08l-.164 1.02c-.045.301-.24 1.186 1.049.645 1.291-.539 6.916-4.078 9.436-6.975C23.176 14.393 24 12.458 24 10.314" />
+  </svg>
+);
